Finalize implementation for send assets action template
